[ios,macos] Partially revert 06af75
Use static_cast<float> to convert from CGfloat when creating
PropertyValue<float> and avoid breaking backwards-compatibility.

+++ platform/darwin/src/MGLPolygon.mm
@@ -44,7 +44,7 @@
     }
 
     mbgl::FillAnnotation annotation { geometry };
-    annotation.opacity = { [delegate alphaForShapeAnnotation:self] };
+    annotation.opacity = { static_cast<float>([delegate alphaForShapeAnnotation:self]) };
     annotation.outlineColor = { [delegate strokeColorForShapeAnnotation:self] };
     annotation.color = { [delegate fillColorForPolygonAnnotation:self] };
